split into solutions and exercises
