--- jax_tutorial/hyperparameters.py
@@ -1,12 +1,0 @@
-# Define parameters that will be constant throughout the tutorial
-# Constants and network parameters
-_dt = 0.01           # Time step (ms)
-_t_max = 100        # Total time of simulation (ms)
-
-# Neuron parameters
-_tau_m = 20.0       # Membrane time constant (ms)
-_V_reset = -65.0    # Reset potential (mV)
-_V_thresh = -50.0   # Threshold potential (mV)
-_R = 1.0            # Membrane resistance (MΩ)
-
-num_simulations = 3
