--- jax_tutorial/profiling_util_functions.py
@@ -1,87 +1,0 @@
-import jax.numpy as jnp
-from jax import grad, jit, vmap
-from jax import random
-
-layer_sizes = [784, 512, 512, 10]
-step_size = 0.01
-num_epochs = 8
-batch_size = 128
-n_targets = 10
-# A helper function to randomly initialize weights and biases
-# for a dense neural network layer
-def random_layer_params(m, n, key, scale=1e-2):
-  w_key, b_key = random.split(key)
-  return scale * random.normal(w_key, (n, m)), scale * random.normal(b_key, (n,))
-
-# Initialize all layers for a fully-connected neural network with sizes "sizes"
-def init_network_params(sizes, key):
-  keys = random.split(key, len(sizes))
-  return [random_layer_params(m, n, k) for m, n, k in zip(sizes[:-1], sizes[1:], keys)]
-
-import numpy as np
-from jax.tree_util import tree_map
-from torch.utils import data
-from torchvision.datasets import MNIST
-
-def numpy_collate(batch):
-  return tree_map(np.asarray, data.default_collate(batch))
-
-class NumpyLoader(data.DataLoader):
-  def __init__(self, dataset, batch_size=1,
-                shuffle=False, sampler=None,
-                batch_sampler=None, num_workers=0,
-                pin_memory=False, drop_last=False,
-                timeout=0, worker_init_fn=None):
-    super(self.__class__, self).__init__(dataset,
-        batch_size=batch_size,
-        shuffle=shuffle,
-        sampler=sampler,
-        batch_sampler=batch_sampler,
-        num_workers=num_workers,
-        collate_fn=numpy_collate,
-        pin_memory=pin_memory,
-        drop_last=drop_last,
-        timeout=timeout,
-        worker_init_fn=worker_init_fn)
-
-class FlattenAndCast(object):
-  def __call__(self, pic):
-    return np.ravel(np.array(pic, dtype=jnp.float32))
-
-
-from jax.scipy.special import logsumexp
-
-def relu(x):
-  return jnp.maximum(0, x)
-
-def predict(params, image):
-  # per-example predictions
-  activations = image
-  for w, b in params[:-1]:
-    outputs = jnp.dot(w, activations) + b
-    activations = relu(outputs)
-  
-  final_w, final_b = params[-1]
-  logits = jnp.dot(final_w, activations) + final_b
-  return logits - logsumexp(logits)
-
-batched_predict = vmap(predict, in_axes=(None, 0))
-
-def one_hot(x, k, dtype=jnp.float32):
-  """Create a one-hot encoding of x of size k."""
-  return jnp.array(x[:, None] == jnp.arange(k), dtype)
-  
-def accuracy(params, images, targets):
-  target_class = jnp.argmax(targets, axis=1)
-  predicted_class = jnp.argmax(batched_predict(params, images), axis=1)
-  return jnp.mean(predicted_class == target_class)
-
-def loss(params, images, targets):
-  preds = batched_predict(params, images)
-  return -jnp.mean(preds * targets)
-
-@jit
-def update(params, x, y):
-  grads = grad(loss)(params, x, y)
-  return [(w - step_size * dw, b - step_size * db)
-          for (w, b), (dw, db) in zip(params, grads)]
